Adjust imports for moved file defaultMutations
