+++ packages/lesswrong/lib/collections/arbitalCache/collection.ts
@@ -1,8 +1,6 @@
 import { createCollection } from '../../vulcan-lib/collections';
-import { getDefaultMutations } from '../../vulcan-core/default_mutations';
 import { ensureIndex } from '../../collectionIndexUtils';
 import { addUniversalFields } from "../../collectionUtils";
-import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
 
 const schema: SchemaType<"ArbitalCaches"> = {
   pageAlias: {
