+++ packages/lesswrong/lib/collections/arbitalTagContentRels/collection.ts
@@ -2,7 +2,7 @@ import { createCollection } from '../../vulcan-lib/collections';
 import schema from './schema';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 
 export const ArbitalTagContentRels = createCollection({
   collectionName: 'ArbitalTagContentRels',
