+++ packages/lesswrong/lib/collections/collections/collection.ts
@@ -3,7 +3,7 @@ import schema from './schema';
 import { makeEditable } from '../../editor/make_editable';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 
 export const Collections: CollectionsCollection = createCollection({
   collectionName: 'Collections',
