+++ packages/lesswrong/lib/collections/electionCandidates/collection.ts
@@ -1,7 +1,7 @@
 import { createCollection } from "../../vulcan-lib/collections";
 import schema from "./schema";
 import { addUniversalFields } from "../../collectionUtils";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
 
 const ElectionCandidates: ElectionCandidatesCollection = createCollection({
