+++ packages/lesswrong/lib/collections/forumEvents/collection.ts
@@ -1,6 +1,6 @@
 import schema from "./schema";
 import { createCollection } from "../../vulcan-lib/collections";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 import { makeEditable } from "../../editor/make_editable";
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
