+++ packages/lesswrong/lib/collections/multiDocuments/collection.ts
@@ -6,7 +6,7 @@ import { canMutateParentDocument, getRootDocument } from "./helpers";
 import { addSlugFields } from "@/lib/utils/schemaUtils";
 import { createCollection } from "@/lib/vulcan-lib/collections.ts";
 import { addUniversalFields } from "@/lib/collectionUtils";
-import { getDefaultMutations } from "@/lib/vulcan-core/default_mutations.ts";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 import { getDefaultResolvers } from "@/lib/vulcan-core/default_resolvers.ts";
 
 export const MultiDocuments = createCollection({
