+++ packages/lesswrong/lib/collections/podcastEpisodes/collection.ts
@@ -3,7 +3,7 @@ import { createCollection } from '../../vulcan-lib/collections';
 import { userIsAdmin, userIsPodcaster } from '../../vulcan-users/permissions';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 
 export const PodcastEpisodes: PodcastEpisodesCollection = createCollection({
   collectionName: 'PodcastEpisodes',
