+++ packages/lesswrong/lib/collections/postViewTimes/collection.ts
@@ -1,5 +1,5 @@
 import { createCollection } from '../../vulcan-lib/collections';
-import { getDefaultMutations } from '../../vulcan-core/default_mutations';
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 import { schema } from './schema';
 import { ensureIndex } from '../../collectionIndexUtils';
 import { addUniversalFields } from "../../collectionUtils";
