+++ packages/lesswrong/lib/collections/reports/collection.ts
@@ -3,7 +3,7 @@ import { userCanDo } from '../../vulcan-users/permissions';
 import { createCollection } from '../../vulcan-lib/collections';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 
 const Reports: ReportsCollection = createCollection({
   collectionName: 'Reports',
