+++ packages/lesswrong/lib/collections/splashArtCoordinates/collection.ts
@@ -2,7 +2,7 @@ import { createCollection } from '../../vulcan-lib/collections';
 import { ensureIndex } from '../../collectionIndexUtils';
 import { schema } from './schema';
 import { userIsAdminOrMod } from '../../vulcan-users/permissions';
-import { getDefaultMutations, MutationOptions } from '../../vulcan-core/default_mutations';
+import { getDefaultMutations, type MutationOptions } from '@/server/resolvers/defaultMutations';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
