+++ packages/lesswrong/lib/collections/subscriptions/collection.ts
@@ -1,7 +1,7 @@
 import schema, { subscriptionTypes } from './schema';
 import { userCanDo, userIsAdmin, userOwns } from '../../vulcan-users/permissions';
 import { createCollection } from '../../vulcan-lib/collections';
-import { getDefaultMutations, MutationOptions } from '../../vulcan-core/default_mutations';
+import { getDefaultMutations, type MutationOptions } from '@/server/resolvers/defaultMutations';
 import Localgroups from '../localgroups/collection';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
