+++ packages/lesswrong/lib/collections/tagRels/collection.ts
@@ -6,7 +6,7 @@ import { isEAForum } from '../../instanceSettings';
 import { userOwns } from '../../vulcan-users/permissions';
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 
 const schema: SchemaType<"TagRels"> = {
   tagId: {
