+++ packages/lesswrong/lib/collections/typingIndicators/collection.ts
@@ -4,7 +4,7 @@ import {userOwns} from "../../vulcan-users/permissions";
 import schema from "./schema";
 import { addUniversalFields } from "../../collectionUtils";
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
-import { getDefaultMutations, MutationOptions } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations, type MutationOptions } from '@/server/resolvers/defaultMutations';
 
 export const TypingIndicators: TypingIndicatorsCollection = createCollection({
   collectionName: 'TypingIndicators',
