+++ packages/lesswrong/lib/collections/userTagRels/collection.ts
@@ -3,7 +3,7 @@ import { foreignKeyField, schemaDefaultValue } from '../../utils/schemaUtils';
 import { createCollection } from '../../vulcan-lib/collections';
 import { userIsAdmin, userOwns } from "../../vulcan-users/permissions";
 import { addUniversalFields } from "../../collectionUtils";
-import { getDefaultMutations } from "../../vulcan-core/default_mutations";
+import { getDefaultMutations } from '@/server/resolvers/defaultMutations';
 import { getDefaultResolvers } from "../../vulcan-core/default_resolvers";
 
 const schema: SchemaType<"UserTagRels"> = {
